Added a message to help debug bad input files
Instead of just saying the following:
FATAL ERROR: End of File (EOF) hit in skipcomments..
Execution Terminated.

when a custom atom's parameters are not found, I added
a message that says it is looking for the custom atom's
parameters right before that message comes up. That can
help people debug their problems with their input files.